pyproject.toml
[project]
name = "adk-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "pypdf>=5.6.0",
]
